move NumberInput example
